+++ src/briefy/common/db/mixins/__init__.py
@@ -9,9 +9,11 @@ from briefy.common.db.mixins.asset import VideoMixin  # noQA
 from briefy.common.db.mixins.external import KnackMixin  # noQA
 from briefy.common.db.mixins.identifiable import GUID
 from briefy.common.db.mixins.metadata import BaseMetadata  # noQA
+from briefy.common.db.mixins.metadata import BaseMetadata  # noQA
 from briefy.common.db.mixins.person import NameMixin  # noQA
 from briefy.common.db.mixins.person import PersonalInfoMixin  # noQA
 from briefy.common.db.mixins.roles import BriefyRoles  # noQA
+from briefy.common.db.mixins.roles import LocalRolesMixin  # noQA
 from briefy.common.db.mixins.timestamp import Timestamp
 from briefy.common.db.mixins.workflow import Workflow