+++ HISTORY.rst
@@ -14,6 +14,7 @@ History
 * Fix identation of code examples in the Permission docstring (ericof).
 * Add PhotoCategoryChoices and VideoCategoryChoices vocabularies (ericof).
 * Add ThreeSixtyImage and Video mixins. Move Image mixin to assets module. (ericof).
+* Add LocalRole support. (ericof).
 
 
 1.1.2 (2016-10-04)